API/requirements.txt
azure-functions

tensorflow-cpu
pydantic
opencensus
opencensus-ext-azure
pytest